requirements.txt
flask
groq
requests
python-dotenv